[codex-security] Split preflight configuration coverage
Move runtime/preflight configuration and provider-projection coverage into a focused API test shard. Keep the monorepo API test below its 150 KB policy limit while preserving identical test coverage and the matching public repository layout.

Linear: FOO-2183
