+++ sdk/typescript/tests-ts/api.test.ts
@@ -38,14 +38,10 @@ import {
   classifyConnectionFailure,
   environmentValue,
   initialCredentialsAvailable,
-  scanPreflightCodexConfig,
-  scanRuntimeCodexConfig,
 } from "../src/api.js";
 import {
   FIREWORKS_CODEX_PROVIDER,
   OPENROUTER_CODEX_PROVIDER,
-  scanModelProvider,
-  writeCodexConfig,
   type JsonObject,
 } from "../src/config.js";
 import { estimateScanCost, type ScanCost } from "../src/cost.js";
@@ -304,353 +300,6 @@ describe("CodexSecurity orchestration", () => {
     expect(environmentValue({ Home: "/shell-home" }, "HOME")).toBe(
       "/shell-home",
     );
-  });
-
-  test("uses a root-read filesystem profile with writable workspace and workbench state", () => {
-    const stateDirectory = join(tmpdir(), "codex-security-persistent-state");
-    const original = {
-      sandbox_mode: "workspace-write",
-      allow_login_shell: true,
-      default_permissions: "unsafe",
-      permissions: {
-        existing: { filesystem: { ":root": "read" } },
-        codex_security_scan: {
-          extends: ":workspace",
-          filesystem: { ":tmpdir": "write" },
-        },
-      },
-    };
-
-    expect(scanRuntimeCodexConfig(original, stateDirectory)).toEqual({
-      allow_login_shell: false,
-      default_permissions: "codex_security_scan",
-      permissions: {
-        existing: { filesystem: { ":root": "read" } },
-        codex_security_scan: {
-          filesystem: {
-            ":root": "read",
-            ":workspace_roots": "write",
-            [stateDirectory]: "write",
-          },
-        },
-      },
-    });
-    expect(original).toMatchObject({
-      sandbox_mode: "workspace-write",
-      allow_login_shell: true,
-      default_permissions: "unsafe",
-    });
-  });
-
-  test("keeps persistent credentials read-only within writable scan state", () => {
-    const stateDirectory = join(tmpdir(), "codex-security-persistent-state");
-    const credentialHome = join(stateDirectory, "codex-home");
-    const config = scanRuntimeCodexConfig({}, stateDirectory, credentialHome);
-
-    expect(config).toMatchObject({
-      permissions: {
-        codex_security_scan: {
-          filesystem: {
-            ":root": "read",
-            ":workspace_roots": "write",
-            [stateDirectory]: "write",
-            [credentialHome]: "read",
-          },
-        },
-      },
-    });
-  });
-
-  test("projects only capability and trust metadata into the readable preflight config", async () => {
-    const root = await temporaryDirectory();
-    const configPath = join(root, "config-preflight.toml");
-    const repository = join(root, "repository");
-    const ordinaryProject = join(
-      root,
-      "settings-service-development-monkey-dataset",
-    );
-    await mkdir(repository);
-    const sanitized = scanPreflightCodexConfig({
-      model: "gpt-5.6-sol",
-      model_reasoning_effort: "high",
-      features: {
-        plugins: true,
-        goals: true,
-        multi_agent_v2: { enabled: false, secret: "FEATURE_SECRET" },
-        api_key: "FEATURE_KEY",
-      },
-      agents: { max_threads: 12, max_depth: 2, token: "AGENT_TOKEN" },
-      profile: "review",
-      profiles: {
-        review: {
-          model: "profile-model",
-          features: { goals: true, secret: "PROFILE_SECRET" },
-          agents: { max_threads: 4, token: "PROFILE_AGENT_TOKEN" },
-          shell_environment_policy: { set: { SECRET: "PROFILE_ENV_SECRET" } },
-        },
-        secret_profile: { features: { goals: false } },
-        "credential-prod": { features: { goals: false } },
-        development: { features: { goals: true } },
-        ["a".repeat(129)]: { features: { goals: false } },
-      },
-      project_root_markers: [
-        ".git",
-        ".workspace",
-        ".env",
-        "PASSWORD_VALUE",
-        "settings.gradle",
-        "a".repeat(257),
-      ],
-      projects: {
-        [repository]: { trust_level: "trusted", token: "PROJECT_TOKEN" },
-        [join(root, "secret-project")]: { trust_level: "trusted" },
-        [join(root, "bearer-PRIVATE")]: { trust_level: "trusted" },
-        [ordinaryProject]: { trust_level: "untrusted" },
-        relative: { trust_level: "trusted" },
-        [join(root, "bad-trust")]: { trust_level: "PROJECT_SECRET" },
-      },
-      mcp_servers: { private: { bearer_token: "MCP_TOKEN" } },
-      shell_environment_policy: { set: { SECRET: "SHELL_SECRET" } },
-    });
-    expect(sanitized).toEqual({
-      model: "gpt-5.6-sol",
-      model_reasoning_effort: "high",
-      features: { goals: true, multi_agent_v2: { enabled: false } },
-      agents: { max_threads: 12, max_depth: 2 },
-      profile: "review",
-      profiles: {
-        review: {
-          model: "profile-model",
-          features: { goals: true },
-          agents: { max_threads: 4 },
-        },
-        development: { features: { goals: true } },
-      },
-      project_root_markers: [".git", ".workspace", "settings.gradle"],
-      projects: {
-        [repository]: { trust_level: "trusted" },
-        [ordinaryProject]: { trust_level: "untrusted" },
-      },
-    });
-    await writeCodexConfig(configPath, sanitized);
-    const serialized = await readFile(configPath, "utf8");
-    for (const secret of [
-      "FEATURE_SECRET",
-      "FEATURE_KEY",
-      "AGENT_TOKEN",
-      "PROFILE_SECRET",
-      "PROFILE_AGENT_TOKEN",
-      "PROFILE_ENV_SECRET",
-      "PROJECT_TOKEN",
-      "MCP_TOKEN",
-      "SHELL_SECRET",
-    ]) {
-      expect(serialized).not.toContain(secret);
-    }
-    const interpreter =
-      Bun.which("python3") ?? Bun.which("python") ?? Bun.which("py");
-    expect(interpreter).not.toBeNull();
-    const output = execFileSync(
-      interpreter!,
-      [
-        join(PLUGIN_ROOT, "scripts", "config_preflight.py"),
-        "--skill",
-        "security-scan",
-        "--config",
-        configPath,
-        "--cwd",
-        repository,
-        "--multi-agent-runtime-owner",
-        "native",
-        "--multi-agent-runtime-version",
-        "v1",
-        "--multi-agent-runtime-provenance",
-        "tool-surface",
-        "--runtime-check",
-        "delegation_available=true",
-        "--runtime-check",
-        "goal_tools_available=true",
-        "--effective-config",
-        "features.goals=true",
-      ],
-      {
-        env: { PATH: process.env["PATH"], CODEX_HOME: join(root, "denied") },
-        encoding: "utf8",
-      },
-    );
-    const preflight = JSON.parse(output) as Record<string, unknown>;
-    expect(preflight["status"]).toBe("ready");
-    expect(preflight["config_resolution"]).toBe("manual-layers");
-    expect(preflight["config_paths"]).toEqual([configPath]);
-    expect(preflight["config_profile"]).toBe("review");
-    expect(JSON.stringify(preflight)).toContain("max_threads");
-    expect(JSON.stringify(preflight)).toContain("12");
-
-    const bridgeConfigPath = join(root, "bridge-preflight.toml");
-    const bridge = scanPreflightCodexConfig({
-      features: { goals: true },
-      multiagent_config: {
-        max_concurrency: 12,
-        token: "BRIDGE_TOKEN",
-      },
-      mcp_servers: { private: { bearer_token: "BRIDGE_MCP_TOKEN" } },
-    });
-    expect(bridge).toEqual({
-      features: { goals: true },
-      multiagent_config: { max_concurrency: 12 },
-    });
-    await writeCodexConfig(bridgeConfigPath, bridge);
-    const bridgeSerialized = await readFile(bridgeConfigPath, "utf8");
-    expect(bridgeSerialized).not.toContain("BRIDGE_TOKEN");
-    expect(bridgeSerialized).not.toContain("BRIDGE_MCP_TOKEN");
-    const bridgeOutput = execFileSync(
-      interpreter!,
-      [
-        join(PLUGIN_ROOT, "scripts", "config_preflight.py"),
-        "--skill",
-        "security-scan",
-        "--config",
-        bridgeConfigPath,
-        "--cwd",
-        repository,
-        "--multi-agent-runtime-owner",
-        "codex-bridge",
-        "--multi-agent-runtime-version",
-        "v2",
-        "--multi-agent-runtime-provenance",
-        "verified-bridge",
-        "--runtime-check",
-        "delegation_available=true",
-        "--runtime-check",
-        "goal_tools_available=true",
-        "--effective-config",
-        "features.goals=true",
-      ],
-      {
-        env: { PATH: process.env["PATH"], CODEX_HOME: join(root, "denied") },
-        encoding: "utf8",
-      },
-    );
-    const bridgePreflight = JSON.parse(bridgeOutput) as Record<string, unknown>;
-    expect(bridgePreflight["status"]).toBe("ready");
-    expect(bridgePreflight["config_resolution"]).toBe("manual-layers");
-    expect(bridgePreflight["config_paths"]).toEqual([bridgeConfigPath]);
-    expect(bridgePreflight["multi_agent_mode"]).toBe("bridge-v2");
-    expect(JSON.stringify(bridgePreflight)).toContain(
-      "multiagent_config.max_concurrency",
-    );
-    expect(JSON.stringify(bridgePreflight)).toContain("12");
-    expect(() =>
-      scanPreflightCodexConfig({
-        projects: Object.fromEntries(
-          Array.from({ length: 256 }, (_, index) => [
-            `/workspace/${index}/${"界".repeat(1300)}`,
-            { trust_level: "trusted" },
-          ]),
-        ),
-      }),
-    ).toThrow(
-      "sanitized Codex Security preflight config exceeds the size limit",
-    );
-    const emptyV2 = scanPreflightCodexConfig({
-      features: {
-        multi_agent_v2: {
-          unknown: true,
-          max_concurrent_threads_per_session: 1_000_001,
-        },
-      },
-    });
-    expect(emptyV2).toEqual({});
-    await expect(
-      writeCodexConfig(join(root, "empty-v2.toml"), emptyV2),
-    ).resolves.toBeUndefined();
-  });
-
-  test("prioritizes the selected profile and active project before projection limits", () => {
-    const activeProject = "/workspace/active";
-    const profiles = Object.fromEntries([
-      ...Array.from({ length: 256 }, (_, index) => [
-        `profile_${index}`,
-        { features: { goals: index % 2 === 0 } },
-      ]),
-      ["selected", { agents: { max_threads: 17 } }],
-    ]);
-    const projects = Object.fromEntries([
-      ...Array.from({ length: 256 }, (_, index) => [
-        `/workspace/project-${index}`,
-        { trust_level: "untrusted" },
-      ]),
-      [activeProject, { trust_level: "trusted" }],
-    ]);
-
-    const prioritized = scanPreflightCodexConfig(
-      {
-        profile: "selected",
-        profiles,
-        projects,
-      },
-      join(activeProject, "packages", "service"),
-    );
-
-    expect(prioritized["profile"]).toBe("selected");
-    expect(Object.keys(prioritized["profiles"] as JsonObject)).toHaveLength(
-      256,
-    );
-    expect(prioritized["profiles"]).toMatchObject({
-      selected: { agents: { max_threads: 17 } },
-    });
-    expect(Object.keys(prioritized["projects"] as JsonObject)).toHaveLength(
-      256,
-    );
-    expect(prioritized["projects"]).toMatchObject({
-      [activeProject]: { trust_level: "trusted" },
-    });
-
-    const validProfiles = Object.fromEntries(
-      Array.from({ length: 256 }, (_, index) => [
-        `valid_${index}`,
-        { features: { goals: true } },
-      ]),
-    );
-    const validProjects = Object.fromEntries(
-      Array.from({ length: 256 }, (_, index) => [
-        `/valid/project-${index}`,
-        { trust_level: "trusted" },
-      ]),
-    );
-    const afterInvalid = scanPreflightCodexConfig({
-      profiles: {
-        ...Object.fromEntries(
-          Array.from({ length: 256 }, (_, index) => [
-            `invalid.profile.${index}`,
-            { features: { goals: false } },
-          ]),
-        ),
-        ...validProfiles,
-      },
-      projects: {
-        ...Object.fromEntries(
-          Array.from({ length: 256 }, (_, index) => [
-            `relative-${index}`,
-            { trust_level: "trusted" },
-          ]),
-        ),
-        ...validProjects,
-      },
-    });
-
-    expect(Object.keys(afterInvalid["profiles"] as JsonObject)).toHaveLength(
-      256,
-    );
-    expect(afterInvalid["profiles"]).toMatchObject({
-      valid_255: { features: { goals: true } },
-    });
-    expect(Object.keys(afterInvalid["projects"] as JsonObject)).toHaveLength(
-      256,
-    );
-    expect(afterInvalid["projects"]).toMatchObject({
-      "/valid/project-255": { trust_level: "trusted" },
-    });
   });
 
   test("selects a real-scan target in the active repository layout", async () => {
@@ -1116,98 +765,6 @@ describe("CodexSecurity orchestration", () => {
       ).toBe(false);
     }
     await client.close();
-  });
-
-  test.each(EXTERNAL_PROVIDER_CASES)(
-    "keeps the %s provider in sanitized scan configuration",
-    (_name, provider, _apiKey, model, providerConfig) => {
-      expect(
-        scanPreflightCodexConfig({
-          model,
-          model_provider: provider,
-          model_providers: {
-            [provider]: providerConfig,
-            private: { api_key: "synthetic-secret" },
-          },
-        }),
-      ).toEqual({
-        model,
-        model_provider: provider,
-        model_providers: { [provider]: providerConfig },
-      });
-    },
-  );
-
-  test.each(EXTERNAL_PROVIDER_CASES)(
-    "preserves the selected %s profile provider in saved scan recipes",
-    (_name, provider, _apiKey, model, providerConfig) => {
-      const config = scanPreflightCodexConfig({
-        model_provider: "openai",
-        profile: "selected",
-        profiles: { selected: { model, model_provider: provider } },
-        model_providers: {
-          [provider]: {
-            ...providerConfig,
-            api_key: "synthetic-provider-secret",
-          },
-          private: { bearer_token: "synthetic-unrelated-secret" },
-        },
-      });
-
-      expect(config).toEqual({
-        model_provider: "openai",
-        profile: "selected",
-        profiles: { selected: { model, model_provider: provider } },
-        model_providers: { [provider]: providerConfig },
-      });
-      expect(scanModelProvider(config)).toBe(provider);
-      expect(JSON.stringify(config)).not.toContain("synthetic-");
-    },
-  );
-
-  test("preserves safe Bedrock provider settings in profile-selected scan recipes", () => {
-    const config = scanPreflightCodexConfig({
-      model_provider: "openai",
-      profile: "bedrock",
-      profiles: {
-        bedrock: {
-          model: "openai.gpt-5.6-luna",
-          model_provider: "amazon-bedrock",
-        },
-      },
-      model_providers: {
-        "amazon-bedrock": {
-          aws: {
-            region: "us-east-2",
-            profile: "security-prod",
-            access_key_id: "synthetic-aws-access-key",
-            secret_access_key: "synthetic-aws-secret-key",
-            session_token: "synthetic-aws-session-token",
-            bearer_token: "synthetic-bedrock-bearer",
-          },
-          api_key: "synthetic-provider-key",
-        },
-        openrouter: { api_key: "synthetic-unrelated-key" },
-      },
-    });
-
-    expect(config).toEqual({
-      model_provider: "openai",
-      profile: "bedrock",
-      profiles: {
-        bedrock: {
-          model: "openai.gpt-5.6-luna",
-          model_provider: "amazon-bedrock",
-        },
-      },
-      model_providers: {
-        "amazon-bedrock": {
-          aws: { region: "us-east-2", profile: "security-prod" },
-        },
-      },
-    });
-    expect(scanModelProvider(config)).toBe("amazon-bedrock");
-    expect(JSON.stringify(config)).not.toContain("synthetic-");
   });
 
   test.each(EXTERNAL_PROVIDER_CASES)(
